// SofTec_TMS/NWC.Web/src/app/TMS-Module/Models/Contract-Station-ListDTO.ts
export class ContractStationListDTO {
       StationID: string;
       branchId: string;
       ContractID: number;
       ContactPersonID: number;
       stationName: string;
       stationCode: string;
       AreaId: string;
       FullName: string;
       Mobile: string;
       FirstName: string;
       SecondName: string;
       LastName: string;
       LandlineNumber: string;
       LandlineNumbeCode: string;
       MobileCode: string;
       Email: string;
       PersonAddress: string;
       PersonAddressPostalCode: string;
       Position: string;
       PersonalIDType: number;
       isDeleted: boolean;
       PersonalIDNumber: string;
       contractStationID: number;
       BranchName: string;
 }